Implement user-defined name hashing.
ANGLEBUG=315
Review URL: https://codereview.appspot.com/6818109

git-svn-id: https://angleproject.googlecode.com/svn/branches/dx11proto@1469 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ include/GLSLANG/ShaderLang.h
@@ -167,7 +167,7 @@ COMPILER_EXPORT int ShFinalize();
 
 // The 64 bits hash function. The first parameter is the input string; the
 // second parameter is the string length.
-typedef khronos_uint64_t (*ShHashFunction64)(const char*, size_t);
+typedef khronos_uint64_t (*ShHashFunction64)(const char*, unsigned int);
 
 //
 // Implementation dependent built-in resources (constants and extensions).
